Crate virus

Ce fichier contient le code source du client CàD le virus lui-même, qui recevras des commandes que l’on lui envoie afin de les executer sur l’hote distant

Son fonctionnement se résume de la manière suivante: Il se connecte au serveur, récupère les commandes envoyées, les execute sur la machine sur laquelle il se trouve, puis renvoie tout message d’execution de la commande. Elle attendras par la suite un nouveau message.

Functions

establish_connection

Cette fonction établie la connection avec le serveur.

execute_command

Cette fonction récupère le parametre command et l’execute.

receive_message

Cette fonction reçois les messages envoyés par le serveur, et les envoie à execute_command